rename some files for organizing
