+++ src/index.css
@@ -11,3 +11,35 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
     monospace;
 }
+
+body,
+p,
+img,
+div,
+h1,
+h2,
+h3,
+h4,
+h5,
+h6,
+a {
+  margin: 0;
+  padding: 0;
+  text-decoration: none;
+  font-family: "Raleway", sans-serif;
+  font-weight: 300;
+}
+
+body {
+  font-family: "Raleway", sans-serif;
+  background-color: #ffffff;
+  /* background-color: #dcdcdd; */
+}
+
+h1, h2, h3, h4, h5, h6, a, p {
+  color: #3f3f3f;
+}
+
+a:visited {
+  color: #46494c;
+}
