+++ src/pages/home/home.css
@@ -1,0 +1,6 @@
+#self-pic {
+  margin: 24px;
+  width: 20vw;
+  height: auto;
+  border-radius: 50%;
+}
